requirements.txt
aiogram>=3.4.0
aiohttp
python-dotenv
Pillow
